Add awake session extension handler

+++ components/sdp/sleep/sleep.h
@@ -8,9 +8,9 @@
 
 void goto_sleep_for_microseconds(uint64_t microsecs);
 
-int get_last_sleep_time();
-int get_time_since_start();
-int get_total_time_awake();
+uint64_t get_last_sleep_time();
+uint64_t get_time_since_start();
+uint64_t get_total_time_awake();
 bool is_first_boot();
 int get_sleep_count();
